chore : 7주차 미션 수정

+++ alex/week5/umc9th/src/main/java/com/example/umc9th/domain/review/entity/Review.java
@@ -21,7 +21,7 @@ public class Review extends BaseTime {
     @ManyToOne(fetch=FetchType.LAZY) @JoinColumn(name="store_id", nullable=false)
     private Store store;
 
-    @Column(nullable=false) private Integer star;
+    @Column(nullable=false) private Float star;
     @Column(nullable=false, length=1000) private String content;
     @Column(nullable=false) private Boolean dbStatus;
